+++ tgbotapi.core/src/commonMain/kotlin/dev/inmo/tgbotapi/types/MessageGenerationStopped.kt
@@ -1,0 +1,18 @@
+package dev.inmo.tgbotapi.types
+
+import dev.inmo.tgbotapi.types.chat.PreviewChat
+import kotlinx.serialization.SerialName
+import kotlinx.serialization.Serializable
+
+/**
+ * Information about a user stopping message generation.
+ */
+@Serializable
+data class MessageGenerationStopped(
+    @SerialName(chatField)
+    val chat: PreviewChat,
+    @SerialName(messageThreadIdField)
+    val messageThreadId: MessageThreadId? = null,
+    @SerialName(draftIdField)
+    val draftId: DraftId
+)